+++ lib/sink/ffmpegsink/ffmpeg_sink.go
@@ -22,7 +22,7 @@ type FFmpegSink struct {
 	stderr   io.ReadCloser
 	stdin    io.WriteCloser
 	frames   layer.FrameForwarder
-	rate     int
+	rate     float64
 }
 
 func New(name string, cfg *config.FFmpegSinkCfg, frameCfg *encdec.FrameCfg, alloc encdec.FrameAllocator) *FFmpegSink {
@@ -59,7 +59,7 @@ func (f *FFmpegSink) setupCmd() error {
 	f.cmd.Env = append(f.cmd.Env, fmt.Sprintf("WIDTH=%d", f.Frames().Width))
 	f.cmd.Env = append(f.cmd.Env, fmt.Sprintf("HEIGHT=%d", f.Frames().Height))
 	f.cmd.Env = append(f.cmd.Env, fmt.Sprintf("SIZE=%dx%d", f.Frames().Width, f.Frames().Height))
-	f.cmd.Env = append(f.cmd.Env, fmt.Sprintf("RATE=%d", f.rate))
+	f.cmd.Env = append(f.cmd.Env, fmt.Sprintf("RATE=%f", f.rate))
 	f.cmd.SysProcAttr = &syscall.SysProcAttr{Pdeathsig: syscall.SIGTERM}
 	var err error
 	f.stdin, err = f.cmd.StdinPipe()
@@ -135,6 +135,6 @@ func (f *FFmpegSink) log(msg string, args ...interface{}) {
 	f.Frames().Log(msg, args...)
 }
 
-func (f *FFmpegSink) SetRate(rate int) {
+func (f *FFmpegSink) SetRate(rate float64) {
 	f.rate = rate
 }
